requirements.txt
opencv-python==4.5.3.56
numpy==1.21.1
dlib==19.22.0
Pillow==8.3.2
streamlit==1.0.0
